Refactor document fixtures
