+++ p2panda-rs/src/test_utils/fixtures/new_fixtures/mod.rs
@@ -1,8 +1,10 @@
+mod document_fixtures;
 mod entry_fixtures;
 mod hash_fixtures;
 mod identity_fixtures;
 mod operation_fixtures;
 
+pub use document_fixtures::*;
 pub use entry_fixtures::*;
 pub use hash_fixtures::*;
 pub use identity_fixtures::*;
